update@2021-05-03-12-07

+++ luci-app-cowbping/root/etc/cowbping.sh
@@ -3,7 +3,7 @@
 
 log_limit() {
 logsize=`du /tmp/log/cowbping.log 2>/dev/null|awk '{print $1}'`
-[ $logsize -gt 80 ] || return
+[ "$logsize" -gt 80 ] || return
 cat /tmp/log/cowbping.log >> /tmp/log/cowbping.log_
 echo ">"$(date +"%Y-%m-%d %H:%M:%S")" ：日志文件过大，旧的记录已暂时经转移到/tmp/log/cowbping.log_。" > /tmp/log/cowbping.log 2>/dev/null
 }
@@ -25,8 +25,9 @@ wifi down && wifi up 2>/dev/null
 /etc/init.d/network restart
 ;;
 "5")
-command=$(uci get cowbping.cowbping.command 2>/dev/null)
-eval ${command} 2>/dev/null
+kill -9 $(busybox ps -w | grep 'cbp_cmd' | grep -v 'grep' | awk '{print $1}') >/dev/null 2>&1
+[ -s /etc/config/cbp_cmd ] || return
+bash /etc/config/cbp_cmd 2>/dev/null &
 ;;
 "6")
 MAC=`dd if=/dev/urandom bs=1 count=32 2>/dev/null | md5sum | cut -b 0-12 | sed 's/\(..\)/\1:/g; s/.$//'`
@@ -41,7 +42,7 @@ poweroff 2>/dev/null
 esac
 }
 
-echo "-----Author: wulishui , 20201003 v2.0-----" >> /tmp/log/cowbping.log
+echo "-----Author: wulishui , 20190805->20210416-----" >> /tmp/log/cowbping.log
 delaytime=$(uci get cowbping.cowbping.delaytime 2>/dev/null)
 sleep "$delaytime"
 time=$(uci get cowbping.cowbping.time 2>/dev/null)
